+++ geometries/__init__.py
@@ -1,0 +1,7 @@
+import os, sys
+sys.path.append('..')
+
+import rcwa
+
+from .MetaVehicle import *
+from .Triangles import *
